@@ -46,7 +46,7 @@
     "frontend"
   ],
   "dependencies": {
-    "@react-spring/web": "^9.3.1",
+    "@react-spring/web": "^9.6.1",
     "@use-gesture/react": "^10.1.1",
     "@vant/icons": "^1.7.0",
     "@vant/popperjs": "^1.1.0",
@@ -62,14 +62,14 @@
   "devDependencies": {
     "@rancui/cli": "workspace:*",
     "@rancui/linter": "workspace:*",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^13.4.0",
     "@types/jest": "^27.0.3",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
     "@types/react-transition-group": "^4.4.5",
     "gh-pages": "^3.2.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "jest-canvas-mock": "2.5.2"
   },
   "browserslist": [
