@@ -1,9 +1,9 @@
 import React from 'react';
-import { render, cleanup, fireEvent, waitFor } from '@testing-library/react';
+import { render, cleanup, fireEvent } from '@testing-library/react';
 import TestsEvent from '../../../tests/events';
 import { sleep } from '../../../tests/utils';
 import DatePicker from '../DatePicker';
-import { DatePickerProps, DateTimePickerInstance } from '../PropsType';
+import type { DatePickerProps, DateTimePickerInstance } from '../PropsType';
 
 function filter(type: string, options: string[]): string[] {
   const mod = type === 'year' ? 10 : 5;
@@ -18,7 +18,7 @@ describe('DatePicker', () => {
   function createDatePicker(props?: Partial<DatePickerProps>) {
     const pickerRef = React.createRef<DateTimePickerInstance>();
 
-    const { queryByTestId, container, rerender, debug } = render(
+    const { queryByTestId, container, rerender, debug, unmount } = render(
       <DatePicker ref={pickerRef} {...props} />,
     );
 
@@ -28,123 +28,13 @@ describe('DatePicker', () => {
       debug,
       queryByTestId,
       pickerRef,
+      unmount,
     };
   }
 
   afterEach(() => {
     cleanup();
     jest.restoreAllMocks();
-  });
-
-  // it('default type', async () => {
-  //   const onConfirm = jest.fn();
-  //   const date = new Date(2020, 10, 1, 0, 0);
-
-  //   const { container } = createDatePicker({
-  //     value: date,
-  //     minDate: new Date(2020, 0, 1),
-  //     maxDate: new Date(2025, 10, 1),
-  //     onConfirm,
-  //   });
-
-  //   const confirmBtn = container.querySelector('.rc-picker__confirm');
-  //   fireEvent.click(confirmBtn);
-
-  //   expect(onConfirm.mock.calls[0][0].getFullYear()).toEqual(2020);
-
-  //   const columnWrapper = container.querySelector('.rc-picker-column');
-  //   await TestsEvent.triggerDrag(columnWrapper, [0, -100]);
-  //   await sleep(100);
-  //   fireEvent.click(confirmBtn);
-
-  //   expect(onConfirm.mock.calls[1][0].getFullYear()).toEqual(2023);
-  // });
-
-  it('month-day type', async () => {
-    const onConfirm = jest.fn();
-    const date = new Date(2020, 10, 1, 0, 0);
-
-    const { container } = createDatePicker({
-      type: 'month-day',
-      value: date,
-      minDate: new Date(2020, 0, 1),
-      maxDate: new Date(2025, 10, 1),
-      swipeDuration: 50,
-      onConfirm,
-    });
-
-    await sleep(400);
-
-    const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
-
-    expect(onConfirm.mock.calls[0][0].getMonth()).toEqual(10);
-    expect(onConfirm.mock.calls[0][0].getDate()).toEqual(1);
-
-    await TestsEvent.triggerDrag(container.querySelectorAll('.rc-picker-column')[0], [0, -800]);
-    await sleep(400);
-
-    fireEvent.click(confirmBtn);
-    expect(onConfirm.mock.calls[1][0].getMonth()).toEqual(11);
-    expect(onConfirm.mock.calls[1][0].getDate()).toEqual(1);
-  });
-
-  it('year-month type', async () => {
-    const onConfirm = jest.fn();
-    const date = new Date(2020, 10, 1, 0, 0);
-
-    const { container } = createDatePicker({
-      type: 'year-month',
-      value: date,
-      minDate: new Date(2020, 0, 1),
-      maxDate: new Date(2025, 10, 1),
-      swipeDuration: 50,
-      onConfirm,
-    });
-
-    await sleep(100);
-
-    const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
-    expect(onConfirm.mock.calls[0][0].getFullYear()).toEqual(2020);
-    expect(onConfirm.mock.calls[0][0].getMonth()).toEqual(10);
-
-    await TestsEvent.triggerDrag(container.querySelectorAll('.rc-picker-column')[0], [0, -300]);
-    await sleep(100);
-    await TestsEvent.triggerDrag(container.querySelectorAll('.rc-picker-column')[1], [0, -300]);
-    await sleep(100);
-    fireEvent.click(confirmBtn);
-    expect(onConfirm.mock.calls[1][0].getFullYear()).toEqual(2025);
-    expect(onConfirm.mock.calls[1][0].getMonth()).toEqual(10);
-  });
-
-  it('datehour type', async () => {
-    const onConfirm = jest.fn();
-    const date = new Date(2020, 10, 1, 0, 0);
-
-    const { container } = createDatePicker({
-      type: 'datehour',
-      value: date,
-      minDate: new Date(2020, 0, 1),
-      maxDate: new Date(2025, 10, 1),
-      swipeDuration: 50,
-      onConfirm,
-    });
-
-    const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
-
-    expect(onConfirm.mock.calls[0][0].getHours()).toEqual(0);
-
-    const columnWrapper = container.querySelectorAll('.rc-picker-column')[3];
-    await waitFor(() => {
-      TestsEvent.triggerDrag(columnWrapper, [0, -800]);
-    });
-    await waitFor(() => {
-      fireEvent.click(confirmBtn);
-    });
-
-    expect(onConfirm.mock.calls[1][0].getHours()).toEqual(23);
   });
 
   it('value has an initial value', async () => {
@@ -157,10 +47,12 @@ describe('DatePicker', () => {
       onConfirm,
     });
 
-    await sleep(400);
+    await sleep(200);
 
     const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
+    if (confirmBtn) {
+      await fireEvent.click(confirmBtn);
+    }
 
     expect(onConfirm.mock.calls[0][0]).toEqual(date);
   });
@@ -174,8 +66,12 @@ describe('DatePicker', () => {
       onConfirm,
     });
 
+    await sleep(200);
+
     const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
+    if (confirmBtn) {
+      await fireEvent.click(confirmBtn);
+    }
 
     expect(onConfirm.mock.calls[0][0]).toEqual(minDate);
   });
@@ -192,7 +88,9 @@ describe('DatePicker', () => {
     await sleep(400);
 
     const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
+    if (confirmBtn) {
+      await fireEvent.click(confirmBtn);
+    }
 
     expect(onConfirm.mock.calls[0][0]).toEqual(defaultValue);
   });
@@ -232,7 +130,7 @@ describe('DatePicker', () => {
     await TestsEvent.triggerDrag(columnWrapper, [0, -100]);
     await sleep(100);
 
-    expect(pickerRef.current.getPicker().getValues()).toEqual([
+    expect(pickerRef.current?.getPicker().getValues()).toEqual([
       '2020 year',
       '05 month',
       '05 day',
@@ -249,7 +147,9 @@ describe('DatePicker', () => {
     });
 
     const confirmBtn = container.querySelector('.rc-picker__cancel');
-    fireEvent.click(confirmBtn);
+    if (confirmBtn) {
+      await fireEvent.click(confirmBtn);
+    }
 
     expect(onCancel).toHaveBeenCalled();
   });
@@ -265,7 +165,9 @@ describe('DatePicker', () => {
     });
 
     const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
+    if (confirmBtn) {
+      await fireEvent.click(confirmBtn);
+    }
 
     expect(onConfirm.mock.calls[0][0]).toEqual(maxDate);
   });
@@ -281,7 +183,9 @@ describe('DatePicker', () => {
     });
 
     const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
+    if (confirmBtn) {
+      await fireEvent.click(confirmBtn);
+    }
 
     expect(onConfirm.mock.calls[0][0]).toEqual(minDate);
   });
@@ -307,7 +211,9 @@ describe('DatePicker', () => {
 
     await sleep(100);
     const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
+    if (confirmBtn) {
+      await fireEvent.click(confirmBtn);
+    }
 
     expect(onConfirm.mock.calls[0][0]).toEqual(new Date(2030, 0, 0, 0, 30));
   });
@@ -332,7 +238,9 @@ describe('DatePicker', () => {
     await sleep(100);
 
     const confirmBtn = container.querySelector('.rc-picker__confirm');
-    fireEvent.click(confirmBtn);
+    if (confirmBtn) {
+      await fireEvent.click(confirmBtn);
+    }
 
     await sleep(100);
 
