@@ -1,12 +1,11 @@
 import React, { useEffect, useState } from 'react';
 
-import ReactDOM from 'react-dom';
-
 import { noop } from '../utils';
 import { CloseParams, ImagePreviewProps, ImagePreviewStatic } from './PropsType';
 
 import BaseImagePreview from './ImagePreview';
 import { resolveContainer } from '../utils/dom/getContainer';
+import { unmount, render } from '../utils/dom/render';
 
 const ImagePreview = BaseImagePreview as ImagePreviewStatic;
 
@@ -48,7 +47,7 @@ ImagePreview.open = (props: ImagePreviewProps) => {
       if ((await beforeClose?.(0)) !== false) {
         destroy(p);
 
-        const unmountResult = ReactDOM.unmountComponentAtNode(container);
+        const unmountResult = unmount(container);
         if (unmountResult && container.parentNode) {
           container.parentNode.removeChild(container);
         }
@@ -68,7 +67,7 @@ ImagePreview.open = (props: ImagePreviewProps) => {
       />
     );
   };
-  ReactDOM.render(<TempDialog />, container);
+  render(<TempDialog />, container);
 
   return {
     close: destroy,
