@@ -27,7 +27,7 @@ export interface TabbarProps<T = number | string> extends BaseTypeProps {
   onChange?: (name: T) => void;
 }
 
-export interface TabbarItemProps<T = number | string> extends BaseTypeProps {
+export interface TabbarItemProps<T = number | string> extends Omit<BaseTypeProps, 'children'> {
   name?: T;
   icon?: string | ((active: boolean) => React.ReactNode);
   iconPrefix?: string;
